web_crawler/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
basicauth==0.4.1
black==20.8b1
cffi==1.14.3
click==7.1.2
colorama==0.4.3
cryptography==3.1.1
Django==3.0
django-cors-headers==3.5.0
django-environ==0.4.5
djangorestframework==3.12.1
isort==5.5.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.0
pep8==1.7.1
Pillow==7.2.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.1
pytz==2020.1
regex==2020.9.27
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
wrapt==1.12.1
